Moved hooks into own folder
