+++ src/routes/safe/components/Balances/index.jsx
@@ -14,9 +14,9 @@ import Divider from '~/components/layout/Divider'
 import Link from '~/components/layout/Link'
 import Row from '~/components/layout/Row'
 import { SAFELIST_ADDRESS } from '~/routes/routes'
-import { useFetchTokens } from '~/routes/safe/components/Balances/FetchTokens'
 import SendModal from '~/routes/safe/components/Balances/SendModal'
 import DropdownCurrency from '~/routes/safe/components/DropdownCurrency'
+import { useFetchTokens } from '~/routes/safe/container/Hooks/useFetchTokens'
 import { safeFeaturesEnabledSelector, safeParamAddressFromStateSelector } from '~/routes/safe/store/selectors'
 import { history } from '~/store'
 import { wrapInSuspense } from '~/utils/wrapInSuspense'
